+++ android/app/src/main/AndroidManifest.xml
@@ -10,10 +10,6 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".views.ShipMarketActivity"
-            android:label="@string/title_activity_ship_market"
-            android:theme="@style/AppTheme.NoActionBar"></activity>
-        <activity
             android:name=".views.ShipActivity"
             android:label="@string/title_activity_ship"
             android:theme="@style/AppTheme.NoActionBar" />
