+++ android/app/src/main/java/edu/gatech/cs2340/rajsjarfiles/spacetrader/views/GameActivity.java
@@ -51,7 +51,7 @@ public class GameActivity extends AppCompatActivity {
     }
 
     public void openMarket(View view) {
-        Intent intent = new Intent(view.getContext(), ShipMarketActivity.class);
+        Intent intent = new Intent(view.getContext(), MarketActivity.class);
         startActivity(intent);
     }
